+++ CHANGELOG.md
@@ -29,6 +29,8 @@
   of /MaxLen when setting a value for a comb text field
 * [HexaPDF::Type::AcroForm::TextField#field_value=] to check the value against
   /MaxLen
+* [HexaPDF::Layout::TextLayouter#fit] to not use style valign when doing
+  variable width layouting
 * CLI command `hexapdf form` to allow setting array values when using a template
 * CLI command `hexapdf form` to allow setting file select fields